requirements.txt
openai>=1.42.0
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.8.2
email-validator>=2.2.0
PyMuPDF>=1.24.9
pillow>=10.4.0
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

